backend/requirements.txt
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg[binary]==3.2.12
jira==3.8.0
celery==5.4.0
redis==5.2.1
django-auth-ldap==5.2.0
python-ldap==3.4.5
